MAINT: Moved DailyReturn to protocol module to break circular references
and removed code that solved that same problem with conditional imports.
